# Add center to Stencils.