--- a/src/mapstencil.jl
+++ b/src/mapstencil.jl
@@ -38,12 +38,14 @@ function _return_type(f, A::AbstractStencilArray{<:Any,<:Any,T}, args...) where 
     bc = boundary(A)
     T1 = bc isa Remove ? promote_type(T, typeof(padval(bc))) : T
     emptyneighbors = _zero_values(T1, st)
-    H = typeof(rebuild(st, emptyneighbors, zero(T1)))
+    emptycenters = _zero_center(T1, st)
+    H = typeof(rebuild(st, emptyneighbors, emptycenters))
     # Use nasty broadcast mechanism `_return_type` to get the new eltype
     return Base._return_type(f, Tuple{H,map(eltype, args)...})
 end
 
 _zero_values(::Type{T}, ::Stencil{<:Any,<:Any,L}) where {T,L} = SVector{L,T}(ntuple(_ -> zero(T), L))
+_zero_center(::Type{T}, ::Stencil) where T = zero(T)
 
 kernel_setup() = KernelAbstractions.CPU(; static=true), 64