--- a/src/stencil.jl
+++ b/src/stencil.jl
@@ -231,21 +231,21 @@ macro stencil(name, description)
 
     struct_expr = quote
         struct $name{R,N,L,T} <: Stencil{R,N,L,T}
-            center::T
             neighbors::SVector{L,T}
-            $name{R,N,L,T}(center::T, neighbors::StaticVector{L,T}) where {R,N,L,T} = new{R,N,L,T}(center, neighbors)
+            center::T
+            $name{R,N,L,T}(neighbors::StaticVector{L,T}, center::T) where {R,N,L,T} = new{R,N,L,T}(neighbors,center)
         end
     end
     func_exprs = quote
         
         # Filled stencils
-        $name{R,N,L}(center::T, neighbors::StaticVector{L,T}) where {R,N,L,T} = $name{R,N,L,T}(center, neighbors)
-        $name{R,N}(center::T, neighbors::StaticVector{L,T}) where {R,N,L,T} = $name{R,N,L,T}(center, neighbors)
-        $name{R}(center, args::StaticVector) where R = $name{R,2}(center, args)
-        $name(center, args::StaticVector; radius=1, ndims=2) = $name{radius,ndims}(center, args)
+        $name{R,N,L}(neighbors::StaticVector{L,T}, center::T) where {R,N,L,T} = $name{R,N,L,T}(neighbors,center)
+        $name{R,N}(neighbors::StaticVector{L,T}, center::T) where {R,N,L,T} = $name{R,N,L,T}(neighbors,center)
+        $name{R}(args::StaticVector, center) where R = $name{R,2}(args, center)
+        $name(args::StaticVector, center; radius=1, ndims=2) = $name{radius,ndims}(args, center)
 
         # Empty stencils
-        $name{R,N,L}() where {R,N,L} = $name{R,N,L}(nothing, SVector(ntuple(_ -> nothing, L)))
+        $name{R,N,L}() where {R,N,L} = $name{R,N,L}(SVector(ntuple(_ -> nothing, L)), nothing)
         function $name{R,N}() where {R,N}
             L = length(offsets($name{R,N}))
             $name{R,N,L}()
@@ -254,7 +254,7 @@ macro stencil(name, description)
         $name(; radius=1, ndims=2) = $name{radius,ndims}()
         $name(radius::Int, ndims::Int=2) = $name{radius,ndims}()
 
-        @inline Stencils.rebuild(n::$name{R,N,L}, center, neighbors) where {R,N,L} = $name{R,N,L}(center, neighbors)
+        @inline Stencils.rebuild(n::$name{R,N,L}, neighbors, center) where {R,N,L} = $name{R,N,L}(neighbors, center)
     end
     return Expr(:block, :(Base.@doc $docstring $struct_expr), func_exprs)
 end