--- a/src/stencils/named.jl
+++ b/src/stencils/named.jl
@@ -41,17 +41,17 @@ and the dimensionality `N` of the stencil is taken from the length of
 the first coordinate, e.g. `1`, `2` or `3`.
 """
 struct NamedStencil{K,O,R,N,L,T} <: Stencil{R,N,L,T}
-    center::T
     neighbors::SVector{L,T}
-    function NamedStencil{K,O,R,N,L,T}(center::T, neighbors::SVector{L,T}) where {K,O,R,N,L,T}
+    center::T
+    function NamedStencil{K,O,R,N,L,T}(neighbors::SVector{L,T},center::T) where {K,O,R,N,L,T}
         length(K) == length(O) == L || throw(ArgumentError("Length of keys must match length of offsets and L parameter"))
-        new{K,O,R,N,L,T}(center, neighbors)
+        new{K,O,R,N,L,T}(neighbors, center)
     end
 end
-NamedStencil{K,O,R,N,L}(center::T, neighbors::SVector{L,T}) where {K,O,R,N,L,T} =
-    NamedStencil{K,O,R,N,L,T}(center, neighbors)
+NamedStencil{K,O,R,N,L}(neighbors::SVector{L,T},center::T) where {K,O,R,N,L,T} =
+    NamedStencil{K,O,R,N,L,T}(neighbors, center)
 NamedStencil{K,O,R,N,L}() where {K,O,R,N,L} =
-    NamedStencil{K,O,R,N,L}(nothing, SVector(ntuple(_ -> nothing, L)))
+    NamedStencil{K,O,R,N,L}(SVector(ntuple(_ -> nothing, L)), nothing)
 function NamedStencil{K,O}() where {K,O}
     N = length(first(O))
     R = _positional_radii(N, O)
@@ -78,6 +78,6 @@ end
 
 offsets(::Type{<:NamedStencil{<:Any,O}}) where O = SVector(O)
 
-@inline function rebuild(::NamedStencil{K,O,R,N,L}, center, neighbors) where {K,O,R,N,L}
-    NamedStencil{K,O,R,N,L}(center, neighbors)
+@inline function rebuild(::NamedStencil{K,O,R,N,L}, neighbors, center) where {K,O,R,N,L}
+    NamedStencil{K,O,R,N,L}(neighbors, center)
 end