--- a/src/stencils/rectangle.jl
+++ b/src/stencils/rectangle.jl
@@ -11,16 +11,17 @@ specified with pulles of offsets around the center point,
 one for each dimension.
 """
 struct Rectangle{O,R,N,L,T} <: Stencil{R,N,L,T}
+    center::T
     neighbors::SVector{L,T}
-    function Rectangle{O,R,N,L,T}(neighbors::SVector{L,T}) where {O,R,N,L,T} 
-        new{O,R,N,L,T}(neighbors)
+    function Rectangle{O,R,N,L,T}(center::T, neighbors::SVector{L,T}) where {O,R,N,L,T} 
+        new{O,R,N,L,T}(center, neighbors)
     end
 end
-Rectangle{O,R,N,L}(neighbors::SVector{L,T}) where {O,R,N,L,T} = 
-    Rectangle{O,R,N,L,T}(neighbors)
+Rectangle{O,R,N,L}(center::T, neighbors::SVector{L,T}) where {O,R,N,L,T} = 
+    Rectangle{O,R,N,L,T}(center, neighbors)
 Rectangle{O,R,N,L}() where {O,R,N,L} = 
-    Rectangle{O,R,N,L}(SVector(ntuple(_ -> nothing, L)))
-function Rectangle{O}(args::SVector...) where O
+    Rectangle{O,R,N,L}(nothing, SVector(ntuple(_ -> nothing, L)))
+function Rectangle{O}(center, args::SVector) where O
     all(map(o -> length(o) == 2, O)) ||
         throw(ArgumentError("All offset tuples must have length `2`, got $O"))
     N = length(O)
@@ -28,7 +29,17 @@ function Rectangle{O}(args::SVector...) where O
         max(map(abs, o)...)
     end
     L = prod(length ∘ splat(:), O)
-    Rectangle{O,R,N,L}(args...)
+    Rectangle{O,R,N,L}(center, args)
+end
+function Rectangle{O}() where O
+    all(map(o -> length(o) == 2, O)) ||
+        throw(ArgumentError("All offset tuples must have length `2`, got $O"))
+    N = length(O)
+    R = maximum(O) do o
+        max(map(abs, o)...)
+    end
+    L = prod(length ∘ splat(:), O)
+    Rectangle{O,R,N,L}()
 end
 Rectangle(os1::AxisOffsets, offsets::AxisOffsets...) = Rectangle((os1, offsets...))
 Rectangle(offsets::CustomOffsets) = Rectangle{offsets}()
@@ -40,6 +51,6 @@ function ConstructionBase.constructorof(::Type{Rectangle{O,R,N,L,T}}) where {O,R
     Rectangle{O,R,N,L}
 end
 
-@inline function rebuild(n::Rectangle{O,R,N,L}, neighbors) where {O,R,N,L}
-    Rectangle{O,R,N,L}(neighbors)
+@inline function rebuild(::Rectangle{O,R,N,L}, center, neighbors) where {O,R,N,L}
+    Rectangle{O,R,N,L}(center, neighbors)
 end