Fixes #57.

I had to split the stencil constructors to prevent ambiguity errors with StaticArrays. I do see that I miss some constructors now for nested arrays, the interface is quite large.

Currently I run into an error with the multi nested layers, I could use some help with that, even Cthulhu is not helping me. Could be a wrong constructor somewhere.

```julia
julia> mapstencil(A) do l
           sum(l.l1.b) - sum(l.l2.a)
       end
ERROR: ArgumentError: cannot convert a value to Union{} for assignment
Stacktrace:
  [1] convert(T::Type{Union{}}, x::Int64)
    @ Base ./essentials.jl:455
  [2] setindex!
    @ ./array.jl:994 [inlined]
  [3] setindex!
    @ ./multidimensional.jl:704 [inlined]
  [4] macro expansion
    @ ~/code/Stencils.jl/src/mapstencil.jl:112 [inlined]
  [5] cpu_mapstencil_kerneln0!
    @ ~/.julia/packages/KernelAbstractions/0r40T/src/macros.jl:293 [inlined]
  [6] cpu_mapstencil_kerneln0!(__ctx__::KernelAbstractions.CompilerMetadata{…}, f::var"#11#12", dest::Matrix{…}, source::StencilArray{…})
    @ Stencils ./none:0
  [7] __thread_run(tid::Int64, len::Int64, rem::Int64, obj::KernelAbstractions.Kernel{…}, ndrange::Nothing, iterspace::KernelAbstractions.NDIteration.NDRange{…}, args::Tuple{…}, dynamic::KernelAbstractions.NDIteration.DynamicCheck)
    @ KernelAbstractions ~/.julia/packages/KernelAbstractions/0r40T/src/cpu.jl:142
  [8] __run(obj::KernelAbstractions.Kernel{…}, ndrange::Nothing, iterspace::KernelAbstractions.NDIteration.NDRange{…}, args::Tuple{…}, dynamic::KernelAbstractions.NDIteration.DynamicCheck, static_threads::Bool)
    @ KernelAbstractions ~/.julia/packages/KernelAbstractions/0r40T/src/cpu.jl:109
  [9] (::KernelAbstractions.Kernel{…})(::Function, ::Vararg{…}; ndrange::Nothing, workgroupsize::Nothing)
    @ KernelAbstractions ~/.julia/packages/KernelAbstractions/0r40T/src/cpu.jl:44
 [10] Kernel
    @ ~/.julia/packages/KernelAbstractions/0r40T/src/cpu.jl:37 [inlined]
 [11] mapstencil!(::Function, ::Matrix{…}, ::StencilArray{…})
    @ Stencils ~/code/Stencils.jl/src/mapstencil.jl:102
 [12] mapstencil(::Function, ::StencilArray{Tuple{…}, 2, Int64, 2, Matrix{…}, Layered{…}, Remove{…}, Conditional})
    @ Stencils ~/code/Stencils.jl/src/mapstencil.jl:18
 [13] top-level scope
    @ REPL[9]:1
Some type information was truncated. Use `show(err)` to see complete types.

```