@rafaqz 2024-12-29
Yes covering static arrays constructor ambiguity is a nightmare.

Can I be a huge pain and ask to swap the argument order to have center last, with a default of `nothing` ? Then this will be non-breaking