@evetion 2024-12-30
I switched it around because some of the constructors  used splatting. I can try again, now that I've removed most of those.

If we want to allow nothing in all constructors, I'd have to discouple the T parameter (and add a new one for center?). Currently all empty constructors are backwards compatible.