@evetion 2024-12-30
I've commented out the failing call described above. It boils down to that with a anonymous function, it can't infer the return type on multilayered stencils as

```julia
    return Base._return_type(f, Tuple{H,map(eltype, args)...})
```
becomes `Union{}`.