@rafaqz 2024-12-30
> If we want to allow nothing in all constructors, I'd have to discouple the T parameter (and add a new one for center?). Currently all empty constructors are backwards compatible.

Ugh right yes maybe a break is better than an extra type parameter