@evetion 2024-12-30
Setup `A = rand(3600,3600); moore = Moore(1)`

Before
```
julia> @benchmark mapstencil(mean, $moore, $A)
BenchmarkTools.Trial: 176 samples with 1 evaluation.
 Range (min … max):  26.678 ms … 68.851 ms  ┊ GC (min … max): 0.00% … 61.85%
 Time  (median):     28.205 ms              ┊ GC (median):    4.75%
 Time  (mean ± σ):   28.422 ms ±  3.124 ms  ┊ GC (mean ± σ):  5.59% ±  4.62%

                             ▁ ▃█▃▃▇▁                          
  ▃▃▃▁▃▁▁▁▁▃▃▃▁▁▃▁▄▃▃▃▅▅▅██▇▇██████████▇▆▄▃▁▁▃▁▁▁▁▁▃▁▁▁▃▁▁▃▁▃ ▃
  26.7 ms         Histogram: frequency by time        29.6 ms <

 Memory estimate: 98.88 MiB, allocs estimate: 76.
```
PR:
```
julia> @benchmark mapstencil(mean, $moore, $A)
BenchmarkTools.Trial: 159 samples with 1 evaluation.
 Range (min … max):  29.327 ms … 73.141 ms  ┊ GC (min … max): 0.00% … 60.09%
 Time  (median):     31.356 ms              ┊ GC (median):    4.39%
 Time  (mean ± σ):   31.520 ms ±  3.456 ms  ┊ GC (mean ± σ):  5.29% ±  4.73%

                       ▁▁             ▂  ▁█▂                   
  ▃▁▃▁▁▁▃▅▁▃▃▃▅▃▄▃██▆▃▇██▄▇▅▄▄▃▃▄▆▅▆▄▄█▆▇███▇▄▅▅▁▄▃▁▃▁▅▁▄▃▃▃▃ ▃
  29.3 ms         Histogram: frequency by time        32.9 ms <

 Memory estimate: 98.88 MiB, allocs estimate: 76.
```

11% slower, which is very close to extra center access for Moore{1} `8+1/8 = 1.125`.

I will add some tests, but otherwise this is good to go?