@rafaqz 2024-12-30
Huh maybe we can make it optional to save the time. I was hoping the compiler could see it wasn't used but maybe not with array access