@evetion 2024-12-30
That would complicate the code a bit, not only in terms of constructors, but the optional would require a new type parameter. Also, I'd like to get this in (and Geomorphometry depending on it released) for JuliaEO. But it's your package.