@evetion 2024-12-30
Tinkering with it, you can cheat using the following patch, by not setting the center field, and not using it, for stencils that already include the center. It does does improve a ms for Window and similar stencils that already include the center. But frankly, not sure if it's worth it, given the complexity, and the fact the benchmark noise is around 10%...

```patch
diff --git a/src/array.jl b/src/array.jl
index 09759e3..2c6a5dd 100644
--- a/src/array.jl
+++ b/src/array.jl
@@ -29,7 +29,18 @@ Base.@propagate_inbounds stencil(hood::StencilOrLayered, A::AbstractStencilArray
     rebuild(hood, neighbors(hood, A, I), _center(hood, A, I))
 Base.@propagate_inbounds stencil(A::AbstractStencilArray) = A.stencil

-@inline _center(::Stencil, A::AbstractStencilArray, I::CartesianIndex) = A[I]
+@inline @generated function _center(hood::Stencil, A::AbstractStencilArray{<:Any,<:Any,T}, I::CartesianIndex) where {T}
+    for (I, O) in enumerate(offsets(hood))
+        if sum(abs.(O)) == 0
+            return :(_zero(T))
+        end
+    end
+    return :(A[I])
+end
+
+@inline _zero(::Type{Nothing}) = nothing
+@inline _zero(T) = zero(T)
+
 @inline _center(hood::Layered, A::AbstractStencilArray, I::CartesianIndex) = map(l -> _center(l, A, I), hood)

 """
diff --git a/src/stencil.jl b/src/stencil.jl
index 1e1dc26..129ce7b 100644
--- a/src/stencil.jl
+++ b/src/stencil.jl
@@ -102,7 +102,14 @@ Base.@propagate_inbounds indexat(hood::Stencil, center, i) = CartesianIndex(offs

 Return the value of the central cell in the stencil.
 """
-center(hood::Stencil) = getfield(hood, :center)
+@generated function center(hood::Stencil)
+    for (I, O) in enumerate(offsets(hood))
+        if sum(abs.(O)) == 0
+            return :(@inbounds neighbors(hood)[$I])
+        end
+    end
+    return :(getfield(hood, :center))
+end

 """
     distances(hood::Stencil)

```