@rafaqz 2024-12-30
Yeah prob not worth it, the compiler will mostly figure it's the same read anyway if it's all inlined