@evetion 2025-01-03
Semantics, but this is not actually breaking if you look at the docstrings. The neighbor, center constructor is private in a sense.